PLANNER-2124: Use WebJar Locator extension to remove version from HTML files
This commit only updates quarkus-school-timetabling. spring-boot-school-timetabling
already uses the spring-boot webjar-locator. quarkus-facility-location uses a CDN,
which requires the full version (if it should be changed to use WebJars, it should
be done in a separate commit). By using a webjar locator, one only need to update
the dependency in the pom.xml for it to be updated everywhere. This make it less
likely updating a dependency will lead to breaking changes or accidentally still
refer to the old version.