Modify quarkus-facility-location to use WebJars
Before, quarkus-facility-location used a CDN. A
CDN does have it merits - reducing load on the server
and increased chance of it being cached on the client.
But those merits are lost for a demo that is not accessed
externally, such as those present in this repository.
Using WebJars allow a user to try out the demo offline
(provided they already downloaded the project dependencies)
and allow the demo to still be usable even if the CDN
provider goes down for any reason.