Remove popper dependency
Popper is needed for Bootstrap tooltips:
https://getbootstrap.com/docs/4.5/components/tooltips/
but those were removed in ba35784974ba10e6a4c51a7432ff9d63bdd4167e.

Note that popper will still appear in webjars because it is a dependency
of the bootstrap webjar. We could even exclude it but the goal of this
commit is to clean up the POM and not declare a dependency for something
that is not used.