Call Center: make isSolving check if status is not NOT_SOLVING
Previously, isSolving check if status is equal to SOLVING_ACTIVE.
This meant SOLVING_SCHEDULED returned false for isSolving. Now
that it check it is not NOT_SOLVING, SOLVING_SCHEDULED return true
for isSolving. This matters for ProblemFactChanges; if they come
in when SOLVING_SCHEDULED, we want to send them to the SolverJob
(since otherwise they won't be registered).