Fix runQuickstartsFromSource.bat
Commit 35912dd2e55bf70fe94604edf3ed87b69121d8bd updated `.sh` version
of the script to run from ".." but ran the `.bat` version of the script from ".", which
leads to "mvnw.cmd" not being found. Because of this, the quickstart launcher does
not build or launch in Windows using runQuickstartsFromSource.bat.

After this fix, the quickstart launcher does build and launch. However,
there is still an issue where the quickstarts do not launch on Windows, complaining
"mvnw" not being found (despite the path in the error existing). This issue would need
to be fixed in the Java files.

In regards to the many cd commands: Windows does not revert the
working directory to it previous state after running commands. So
`mvnw -f ..` changes the working directory back to the parent,
so we need to back into build again. And then an extra cd to get
out of quickstarts-showcase.